requirements.txt
Flask
gunicorn==19.6.0
requests
wikipedia
line-bot-sdk
pandas
googletrans
google-cloud-storage
